requirements.txt
python-telegram-bot==20.3
aiohttp==3.8.6
Flask==2.3.5
